+++ contracts/v2/accounts/test/Implementation_m2.test.sol
@@ -2,28 +2,24 @@ pragma solidity ^0.7.0;
 pragma experimental ABIEncoderV2;
 import "hardhat/console.sol";
 
+
 /**
  * @title InstaAccountV2 Mapping 2.
  * @dev DeFi Smart Account Wallet.
  */
 
 interface DefaultImplementation {
-    function version() external view returns (uint256);
-
-    function isAuth(address) external view returns (bool);
+    function version() external view returns(uint);
+    function isAuth(address) external view returns(bool);
 }
 
 interface IndexInterface {
-    function connectors(uint256 version) external view returns (address);
-
-    function check(uint256 version) external view returns (address);
+    function connectors(uint version) external view returns (address);
+    function check(uint version) external view returns (address);
 }
 
 interface ConnectorsInterface {
-    function isConnectors(string[] calldata connectorNames)
-        external
-        view
-        returns (bool, address[] memory);
+    function isConnectors(string[] calldata connectorNames) external view returns (bool, address[] memory);
 }
 
 interface CheckInterface {
@@ -39,19 +35,15 @@ contract InstaImplementationM2 {
         connectorsM2 = _connectors;
         instaIndex = _instaIndex;
     }
-
-    function decodeEvent(bytes memory response)
-        internal
-        pure
-        returns (string memory _eventCode, bytes memory _eventParams)
-    {
+    
+    function decodeEvent(bytes memory response) internal pure returns (string memory _eventCode, bytes memory _eventParams) {
         (_eventCode, _eventParams) = abi.decode(response, (string, bytes));
     }
 
     event LogCast(
         address indexed origin,
         address indexed sender,
-        uint256 value,
+        uint value,
         string[] targetsNames,
         address[] targets,
         string[] eventNames,
@@ -60,41 +52,28 @@ contract InstaImplementationM2 {
 
     receive() external payable {}
 
-    /**
+     /**
      * @dev Delegate the calls to Connector And this function is ran by cast().
      * @param _target Target to of Connector.
      * @param _data CallData of function in Connector.
-     */
-    function spell(address _target, bytes memory _data)
-        internal
-        returns (bytes memory response)
-    {
+    */
+    function spell(address _target, bytes memory _data) internal returns (bytes memory response) {
         require(_target != address(0), "target-invalid");
         assembly {
-            let succeeded := delegatecall(
-                gas(),
-                _target,
-                add(_data, 0x20),
-                mload(_data),
-                0,
-                0
-            )
+            let succeeded := delegatecall(gas(), _target, add(_data, 0x20), mload(_data), 0, 0)
             let size := returndatasize()
-
+            
             response := mload(0x40)
-            mstore(
-                0x40,
-                add(response, and(add(add(size, 0x20), 0x1f), not(0x1f)))
-            )
+            mstore(0x40, add(response, and(add(add(size, 0x20), 0x1f), not(0x1f))))
             mstore(response, size)
             returndatacopy(add(response, 0x20), 0, size)
 
             switch iszero(succeeded)
-            case 1 {
-                // throw if delegatecall failed
-                returndatacopy(0x00, 0x00, size)
-                revert(0x00, size)
-            }
+                case 1 {
+                    // throw if delegatecall failed
+                    returndatacopy(0x00, 0x00, size)
+                    revert(0x00, size)
+                }
         }
     }
 
@@ -103,46 +82,34 @@ contract InstaImplementationM2 {
      * from Smart Account.
      * @param _targetNames Array of Target(s) to of Connector.
      * @param _datas Array of Calldata(S) of function.
-     */
+    */
     function castWithFlashloan(
         string[] calldata _targetNames,
         bytes[] calldata _datas,
         address _origin
     )
-        external
-        payable
-        returns (
-            bytes32 // Dummy return to fix instaIndex buildWithCast function
-        )
-    {
-        DefaultImplementation defaultImplementation = DefaultImplementation(
-            address(this)
-        );
+    external
+    payable 
+    returns (bytes32) // Dummy return to fix instaIndex buildWithCast function
+    {   
+
+        DefaultImplementation defaultImplementation = DefaultImplementation(address(this));
         uint256 _length = _targetNames.length;
 
-        require(
-            defaultImplementation.isAuth(msg.sender) ||
-                msg.sender == address(instaIndex),
-            "InstaImplementationM1: permission-denied"
-        );
-        require(
-            _length == _datas.length,
-            "InstaImplementationM1: array-length-invalid"
-        );
+        require(defaultImplementation.isAuth(msg.sender) || msg.sender == address(instaIndex), "InstaImplementationM1: permission-denied");
+        require(_length == _datas.length , "InstaImplementationM1: array-length-invalid");
 
         string[] memory eventNames = new string[](_length);
         bytes[] memory eventParams = new bytes[](_length);
 
-        (bool isOk, address[] memory _targets) = ConnectorsInterface(
-            connectorsM2
-        ).isConnectors(_targetNames);
+        (bool isOk, address[] memory _targets) = ConnectorsInterface(connectorsM2).isConnectors(_targetNames);
         require(isOk, "1: not-connector");
-
-        for (uint256 i = 0; i < _targets.length; i++) {
+        
+        for (uint i = 0; i < _targets.length; i++) {
             bytes memory response = spell(_targets[i], _datas[i]);
             (eventNames[i], eventParams[i]) = decodeEvent(response);
         }
-
+        
         emit LogCast(
             _origin,
             msg.sender,
@@ -153,4 +120,5 @@ contract InstaImplementationM2 {
             eventParams
         );
     }
+
 }
